Add files via upload
This is a great beginners template for the Titanic data competition on Kaggle. I tried to make it as readable as possible for the newcomer. Hopefully, someone can take this and use it as a teaching set and potentially improve upon its performance. This is a great introduction to a logistic regression competition.